fix sample
