+++ sample/src/mainScene6.ts
@@ -1,4 +1,4 @@
-import {Label} from "@akashic-extension/akashic-label";
+import {Label, RubyAlign, Fragment} from "@akashic-extension/akashic-label";
 var game = g.game;
 
 module.exports = function() {
@@ -6,7 +6,7 @@ module.exports = function() {
 		game: game,
 		assetIds: ["bmpfont", "bmpfont-glyph", "mplus", "mplus-glyph"]
 	});
-	var rate = game.fps / 2;
+	var rate = game.fps / 3;
 	scene.loaded.add(function() {
 
 		// グリフデータの生成
@@ -47,116 +47,101 @@ module.exports = function() {
 			hint: dhint
 		});
 
-		// 複数行のルビ機能
-		var tlabel0 = new Label({
-			scene: scene,
-			text: "複数行のルビ機能",
-			font: mplusfont,
-			fontSize: 30,
-			width: game.width,
-			textAlign: g.TextAlign.Center
-		});
-		tlabel0.x = 0;
-		scene.append(tlabel0);
-
-
-		// ルビと改行
-		var y0 = 40;
-
-		// ルビの行幅切替
-		var label01 = new Label({
-			scene: scene,
-			text:'ルビのある行と無い行で\rルビによる{"rb": "行間幅設定", "rt": "fixLineGap"}を\r切り替えることができます',
-			font: mplusfont,
-			fontSize: 20,
-			width: game.width,
-			fixLineGap: false,
-			rubyOptions: {rubyFont: bmpfont}
-		});
-		label01.y = y0;
-		label01.touchable = true;
-		label01.update.add(function() {
-			if (game.age % rate === 0) {
-				label01.fixLineGap = !label01.fixLineGap;
-				label01.invalidate();
-			}
-		}, label01);
-		scene.append(label01);
-
-		// ルビ応用
-		var textArray = [].concat('「よろしゅうございます。'.split(/.*?/),
-			'{"rb": "南　　", "rt": "　　　　　　　"}',
-			30,
-			'{"rb": "南十　", "rt": "　　　　　　　"}',
-			30,
-			'{"rb": "南十字", "rt": "　　　　　　　"}',
-			30,
-			'{"rb": "南十字", "rt": "サ　　　　　　"}',
-			30,
-			'{"rb": "南十字", "rt": "サウ　　　　　"}',
-			30,
-			'{"rb": "南十字", "rt": "サウザ　　　　"}',
-			30,
-			'{"rb": "南十字", "rt": "サウザン　　　"}',
-			30,
-			'{"rb": "南十字", "rt": "サウザンク　　"}',
-			30,
-			'{"rb": "南十字", "rt": "サウザンクロ　"}',
-			30,
-			'{"rb": "南十字", "rt": "サウザンクロス"}',
-			'へ着きますのは、次の第三時ころになります。」\r'.split(/.*?/),
-			'{"rb": "車", "rt": "　"}',
-			22,
-			'{"rb": "車", "rt": "し"}',
-			22,
-			'{"rb": "車", "rt": "しゃ"}',
-			'{"rb": "掌", "rt": "　　　"}',
-			24,
-			'{"rb": "掌", "rt": "し　　"}',
-			24,
-			'{"rb": "掌", "rt": "しょ　"}',
-			24,
-			'{"rb": "掌", "rt": "しょう"}',
-			'は紙をジョバンニに渡して向うへ行きました。'.split(/.*?/)
-		);
-		var counter = 0;
-		var mlabel = new Label({
-			scene: scene,
-			text:"",
-			font: mplusfont,
-			fontSize: 20,
-			lineGap: 0,
-			width: 240,
-			fixLineGap: true,
-			rubyOptions: {rubyGap: 0}
-		});
-		mlabel.y = 130;
-		counter = 0;
-		mlabel.textAlign = g.TextAlign.Left;
-		scene.append(mlabel);
-		mlabel.update.add(function() {
-			// 初期化と待機
-			if (counter === textArray.length) {
-				this.text = "";
-				counter = -10;
-				return;
-			}
-			if ( counter < textArray.length && game.age % 2 === 0) {
-				if (counter >= 0){
-					if (typeof textArray[counter] === "number") {
-						var n = textArray[counter];
-						this.text = this.text.substring(0, this.text.length - n);
-						counter += 1;
-						this.text += textArray[counter];
-
-					} else {
-						this.text += textArray[counter];
-					}
-					this.invalidate();
+		var counter = 0;		
+		var text = "「これ」と「それ」と「あれ」と「どれ」と「これ」と「それ」と「あれ」と「どれ」と「これ」と「それ」と「あれ」と「どれ」と「これ」と「それ」と「あれ」と「どれ」";
+		var sampleRule = function (text: Fragment[], index: number) {
+			text = (text as string[]);
+			var joinedText = text.join("");
+			const target = joinedText[index];
+				if (target === "」") {
+					return index + 1;
+				} else if (target === "「") {
+					return index - 1;
+				} else {
+					return index;
 				}
-				counter += 1;
+		}
+		var lblabel = new Label({
+			scene: scene,
+			text: text,
+			font: mplusfont,
+			fontSize: 14,
+			textAlign: g.TextAlign.Left,
+			width: game.width / 4,
+			lineBreak: true,
+			widthAutoAdjust: true,
+			lineBreakRule: sampleRule
+		});
+		lblabel.y = 40;
+		scene.append(lblabel);
+		lblabel.update.add(function() {
+			if (game.age % rate === 0) {
+				this.width = counter % 20 * 10 + 120;
+				counter++;
+				this.invalidate();
 			}
-		}, mlabel);
+		}, lblabel);
+
+		var text = '「{"rt":"これ","rb":"これ"}」と「{"rt":"それ","rb":"それ"}」と「{"rt":"あれ","rb":"あれ"}」と「{"rt":"どれ","rb":"どれ"}」と' +
+			'「{"rt":"これ","rb":"これ"}」と「{"rt":"それ","rb":"それ"}」と「{"rt":"あれ","rb":"あれ"}」と「{"rt":"どれ","rb":"どれ"}」と' +
+			'「{"rt":"これ","rb":"これ"}」と「{"rt":"それ","rb":"それ"}」と「{"rt":"あれ","rb":"あれ"}」と「{"rt":"どれ","rb":"どれ"}」と' +
+			'「{"rt":"これ","rb":"これ"}」と「{"rt":"それ","rb":"それ"}」と「{"rt":"あれ","rb":"あれ"}」と「{"rt":"どれ","rb":"どれ"}」と';
+		var sampleRule = function (fragments: Fragment[], index: number) {
+			/*
+			text = text.map((e) => {
+				if (e instanceof String) {
+					return e;
+				} else {
+					return "〇";
+				}
+			});
+			var joinedText = text.join("");
+			*/
+			const target = fragments[index];
+				if (target === "」") {
+					return index + 1;
+				} else if (target === "「") {
+					return index - 1;
+				} else {
+					return index;
+				}
+		}
+		var lblabel2 = new Label({
+			scene: scene,
+			text: text,
+			font: mplusfont,
+			fontSize: 14,
+			textAlign: g.TextAlign.Left,
+			width: game.width / 4,
+			lineBreak: true,
+			widthAutoAdjust: true,
+			lineBreakRule: sampleRule
+		});
+		lblabel2.y = 160;
+		scene.append(lblabel2);
+		lblabel2.update.add(function() {
+			if (game.age % rate === 0) {
+				this.width = counter % 20 * 10 + 120;
+				counter++;
+				this.invalidate();
+			}
+		}, lblabel2);
+
+		var nlabel = new Label({
+			scene: scene,
+			text: "［次＞＞］",
+			font: mplusfont,
+			fontSize: 20,
+			width: game.width
+		});
+		nlabel.x = 230;
+		nlabel.y = game.height - 20;
+		nlabel.touchable = true;
+		nlabel.pointDown.add(function() {
+			var scene3 = require("mainScene")();
+			game.replaceScene(scene3);
+		}, nlabel);
+		scene.append(nlabel);
 
 		var dlabel = new Label({
 			scene: scene,
@@ -170,11 +155,11 @@ module.exports = function() {
 		dlabel.y = game.height - 20;
 		dlabel.touchable = true;
 		dlabel.pointDown.add(function(){
-			scene.children.forEach((label) => {
-				if (label instanceof Label) {
-					label.font = dfont;
-					label.rubyOptions.rubyFont = dfont;
-					label.invalidate();
+			scene.children.forEach((child: g.E) => {
+				if (child instanceof Label) {
+					child.font = dfont;
+					child.rubyOptions.rubyFont = dfont;
+					child.invalidate();
 				}
 			});
 		}, dlabel);
